Ensure that events are processed when unloading
* Updates the `_handleUnload` function to use a synschronous, blocking
`XMLHttpRequest` to prevent page transitions happening before the
data is sent to the API. This should be considered carefully as it
could slow down transitions for users. `navigator.sendBeacon` would
be a good alternative, but sending JSON  is currently
blocked in Chrome due to CORS considerations
 (see here: https://bugs.chromium.org/p/chromium/issues/detail?id=490015).
`sendBeacon` also does not allow custom headers such as Authorization,
 so we could not send through tokens to identify the user.
* Also queues up events that are fired before the session has been
initialized so that these can have the correct keys attached to them
for processing. This prevents either events being dropped, or errors
in validation from the API due to events missing vital keys.