Allow the browserId and sessionId to be provided by the URL params
When setting the ids, the function now checks the URL params for the browser and sesions IDs. If present, these are used and override any other saved IDs. This allows sessions to be continued across sites that do not have access to the same `localStorage` as was originally assumed when this was written for the Kano 2 App. The other option would be to check for a crossStorage client, so that sites using shared storage would be able to share the IDs that way. I was reluctant to take that approach though, since it binds the tracking behavior too closely to the shareds storage, and these would also be asynchronous operations (not necessarily a problem, but this would complicate the tracking behaviour unnecessarily dealing with synchronous and async flows).

The `_getBrowserId` and `_getSessionId` functions make the `browserId` and `sessionId` available to the main app using this behavior/mixin, so that they can be appended to the URLs when switching between sites.

Also ensures that preInit events are properly added to the queue, and the queue cleared, even if the session is being restarted (ie. when the behavior is being initialized afresh, but the session and browser IDs are coming from the URL query params).